feat(gpu): int residency — Sort/Argsort return GPU-resident indices; argsort->gather stays on-device
Completes the int half of Bit+int residency:
- take_along_dim now reads indices from a FLOAT buffer (int values cast in-kernel), so a resident
  int-index tensor is reused in place instead of re-uploaded. The engine passes the indices via the
  resident-capable GetOrAllocateBuffer (int rides as float, exact for the index range).
- BitonicSortRows split into a buffer-returning core (BitonicSortRowsToBuffers); TensorSort now keeps
  both the sorted VALUES and the INT INDICES resident via FinishGpuOp<T>/FinishGpuOp<int> instead of
  downloading. Median/Kthvalue/Mode/NanMedian/Unique keep the download variant (host scalar/dedup picks).

Net: argsort(x) -> take_along_dim(x, idx) keeps the indices on the GPU end-to-end. Verified vs CPU
(full sort family + the argsort->gather chain), 223 parity cases green, net10.0 + net471, no regressions.

+++ src/AiDotNet.Tensors/Engines/DirectGpu/OpenCL/Kernels/ShapeKernels.cs
@@ -104,12 +104,12 @@ __kernel void index_select(__global const float* input, __global const float* in
 }
 // Gather along one axis (take_along_dim/gather): input viewed [outer, axisIn, inner], indices/output
 // viewed [outer, axisOut, inner]; output[o,j,i] = input[o, indices[o,j,i], i].
-__kernel void take_along_dim(__global const float* input, __global const int* indices, __global float* output,
+__kernel void take_along_dim(__global const float* input, __global const float* indices, __global float* output,
     int outerSize, int axisOut, int innerSize, int axisIn) {
     int idx = get_global_id(0); int total = outerSize * axisOut * innerSize; if (idx >= total) return;
     int inner = idx % innerSize;
     int outer = (idx / innerSize) / axisOut;
-    int srcJ = indices[idx];
+    int srcJ = (int)indices[idx];
     if (srcJ < 0 || srcJ >= axisIn) { output[idx] = 0.0f; return; }
     output[idx] = input[(outer * axisIn + srcJ) * innerSize + inner];
 }
